connman client library is in, with tests and all.
during endless nights and talks I debugged it and seems to not leak or
segfault, but now we need a GUI for it!

My idea is to have an gadget much like mixer, it would show the
current connection and status, when you click you'd get a "Controls",
a list of devices and a "offline" checkbox.

Controls would open a detailed control panel with devices (ie:
interfaces, eth0, eth1, wlan0...), networks (access points) and
properties about them, like the passphrase, whenever to remember or
not, autoconnect or not, policy and even reoder (change priority).

List of devices would show them ("Wireless") and possible some
information, like the current access point name or ssid, it's a radio.

Offline checkbox (and e_action) are mandatory so you can easily turn
them off. Connman is smart and will turn off radios, that can also
save power.

Ideas, anyone willing to implement the gadget?




SVN revision: 39807