start service by name requires one more parameter (flags).
The dbus api has a "flags" parameter that is now unused but may be in
future, it was missing and dbus was giving method mismatch.

I forgot to commit this but changed Ethumb_Client, then Viktor "fixed"
it by reverting such change. Now going back to my code and adding "0"
as flag.



SVN revision: 40968