e_dbus: fix 2 --disable for the same thing
PrinceAMD> ahh, guys in e_dbus why is there  --disable-performance-test &
           --disable-performance_test, by default e_dbus says performance
           is disable but still tries to build it and fails, it cant
           find <Eina.h>
PrinceAMD> anyhow with autogen --disable-performance-test
           --disable-performance_test , e_dbus compile

We were actually trying to compile a "component" that is really a test
binary. And for that component we were not applying the CFLAGS. Now
there's only a --disable-performance-test as it was supposed to be.



SVN revision: 74540